[ScfSimpleIfToSelect] Fix bugs and improve performance
This simplifies/improves the logic inside the `ScfSimpleIfToSelect`
transformation pass to increase the amount of situations in which we can
convert an if to a select. In cases where only a single branch contains
a supported arithmetic operation, the pass is now able to hoist that
arithmetic operation without further conditions out of the if and at no
throughput/area cost.

This also fixes a bug where the transformation would yield a semantically
incorrect IR in scenarios where hoisting is now possible. Unit tests are
added to check for correct behavior in such situations.