[integration-test] Fix `gcd` kernel implementation
The implementation of the `gcd` kernel, which computes the GCD of two
signed integers using the "binary GCD" algorithm, was incorrect,
yielding an incorrect result on the test input set in the kernel's main
function. This fixes the mistakes in the algorithm's implementation. The
kernel now returns the correct result for the test input.

Fixes #79.