[HandshakeToHW] Wrapper generation along kernel module
This makes the `HandshakeToHW` conversion pass generate an additional
wrapper module around the hardware module corresponding to the kernel.
The wrapper contains an instance of the kernel module along IO
converters connecting all of the kernel's "ad-hoc" memory interfaces to
dual-port BRAM interfaces exposed in the wrapper's top-level IO. The
corresponding VHDL component implementing the converter is added to the
experimental RTL library, and a corresponding entry for it is created in
the RTL configuration file.