[Backend] Add LSQ generator to build process
Previously, our backend used a pre-built version of the LSQ generator
(a JAR file pulled as-is from legacy Dynamatic) during RTL emission to
create LSQs mathing those in its input circuit. This commit removes this
dependency on legacy Dynamatic and makes Dynamatic build and use its own
version of the same Chisel-based LSQ generator. The project's build
instructions and build process are updated to reflect the change (we now
depend on Scala/Chisel).

The JSON-formatted LSQ configuration file that our LSQ generator expects
is slightly simpler than the legacy one (we no longer require the actual
configuration to be stored in a JSON list of "specifications"). As a
result, both the current `export-vhdl` tool and the `HandshakeToHW`
conversion pass are lightly modified to reflect the change.