[CfToHandshake] Add support for func-level calls
The conversion pass is now able to lower `func::CallOp` operations to
matching `handshake::InstanceOp` operations, effectively adding support
for external function calls to the pass. This was also the opportunity
to slightly refactor the way the conversion pass initially converts
func-level functions to handshake-level functions. This now happens in a
separate greedy rewriting step before the actual.