[HandshakeHoistExtInstances] Hoist instances into IO
This commit adds a new Handshake-level pass meant to inmprove support
for external functions in Dynamatic. The pass replaces instances of
external Handshake functions with additional arguments/results to their
parent Handshake function, effectively "hoisting" them next to the
function (though they are not represented in the IR, the user is
responsible to provide an RTL implementation for them manually).