[export-vhdl] Preparation for Verilog export support
This commit makes architectural changes to the internal data-structured
used by the experimental VHDL export tool to prepare it to be able to
export Verilog as well. The code is restructured to allow maximum reuse
between the current `VHDLWriter` and future `VerilogWriter`. The
exporter's output should not be affected by this change.