[NameAnalysis] Change name attribute format
This commit makes the attribute holding a unique name for an operation
---the one managed by the `NameAnalysis` analysis---follow MLIR's
specification for *discardable attributes*
(https://mlir.llvm.org/docs/LangRef/#attributes).

The attribute was previously named "name" and was of type
`handshake::NameAttr` (a simple wrapper around `StringAttr`); it is now
named "handshake.name" and is of type `StringAttr`. The
`handshake::NameAttr` attribute is no longer necessary and is therefore
removed.