[HandshakeToHW] External instance arguments directly to output
This is a temporary hack to prevent arguments from external Handshake
instances hoisted by the `HandshakeHoistExtInstances` pass from going
through the end synchronizer when lowered to the HW dialect (instead
going directly to the module's outputs). This is achieved using an
ad-hoc attribute on the `handshake.end` terminator that is set by the
hoisting pass and read by the `HandshakeToHW` conversion pass.