[LSQ] Fix bug in LSQ config generation
When the LSQ config generation was moved from the DOT export tool to the
`LSQGenerationInfo` helper struct, a bug was introduced in the config's
`loadOffsets` and `storeOffsets` parameters. While these are
semantically one-dimensional arrays, the legacy LSQ generator expects
them to be bi-dimensional arrays, where only the first value of each
array matters. The `LSQGenerationInfo` struct was representing them as
one-dimensional arrays, yielding to the generation of incorrect LSQs and
simulation failures/deadlocks. This commit makes them bi-dimensional
arrays again, fixing the related problems.