[RTL] Refactor RTL requests for dependencies
This improves the API around RTL requests to allow for a different
matching logic depending on whether the user queries a component
matching an MLIR operation---for which the match is based on the
component's name and parameters---or queries a component that is
the dependency of another one, in which case the match is based on the
concrete module name. The latter is also a change in internal logic;
before this commit dependencies were identified by the component's name
as opposed to the component's module name. The latter makes more sense
w.r.t. the RTL configuration, which doesn't need to define "fake names"
for components only ever serving as dependencies anymore (in these cases
the name can be ommited).

Notably, `RTLRequest` objects heap-allocate `RTLMatche`s when they find
a matching between themselves and a component; these matches then need
to be manually freed.