[Sharing] New Resource Sharing Pass (#77)
This commit introduces a new resource-sharing (one physical unit for
multiple operations) strategy that will be presented at IWLS '24.

It introduces a new resource sharing pass (`CreditBasedSharing`) and a
new `handshake::SharingWrapperOp` operation. 

----

The `SharingWrapperOp` decides which operation to access the shared
resource, and dispatches the computed result to the corresponding
successor.

The following example describes what this pass does to two sharing
targets, `muli0` and `muli1`.

Before sharing:

```mlir
%succOut1 = arith.muli %pred1In1, %pred1In2 { handshake.name="muli0" } : i32
%succOut2 = arith.muli %pred2In1, %pred2In2 { handshake.name="muli1" } : i32
```

After sharing:

```mlir
%sharedOut1 = arith.muli %selIn1, %selIn1 { handshake.name="muliShared1" } : i32
%selIn1, %selIn2, %succOut1, %succOut2 = sharing_wrapper %pred1In1, %pred1In2, %pred2In1, %pred2In2, %sharedOut1 [3, 3] : i32, i32, i32, i32, i32
```

The sharing wrapper selects the original operands (%pred1In1, %pred1In2,
%pred2In1, %pred2In2), and distributes the computed results to the
original results (%succOut1, %succOut2).

It sends the selected results to the shared multiplier through %selIn1
and %selIn2, and takes the result %sharedOut1 of the shared multiplier.

The integer array in the attribute is the number of credits per
operation that shares it, the number of credits sets the achievable
performance of each operation that shares the unit; the sharing pass
decides the credit allocation.

----

This proposed pass (`CreditBasedSharing`) is designed as a drop-in
replacement of the buffer placement pass, i.e., it provides the same
interface as the `HandshakePlaceBuffers`. It internally calls a
derived class from `HandshakePlaceBuffer` and retrieves data from MILP
solutions. The retrieved data are necessary for sustaining the
performance achieved during buffer placement.