[LSQ] Add LSQ version working w/ experimental backend
This adds a version of the BRAM-connected LSQ which works with the
experimental backend. It differs from the original implementation in
three significant ways.

1. Signals always set to constants in the legacy backend are removed
   (the constants become part of the module's RTL description instead).
2. Ports have names that match our naming conventions for memory
   interfaces.
3. Control-only-ports have the proper RTL types (valid and ready wires
   without data).

To support switching between the legacy and experimental implementations
of the LSQ, the LSQ JSON config file expects a new "experimental"
boolean-typed field to indicate to the generator which implementation to
pick.