[Backend] Correct MC/LSQ iface-to-iface connectivity
The experimental backend was generating incorrect RTL in situations
where the input dataflow circuit contained an LSQ connected to a MC
(instead of a direct connection to a BRAM). In such cases, legacy's
behavior was to repurpose the LSQ's non-elastic BRAM-like ports into a
pair of "elastic" load/store ports connecting to properly elastic
load/store ports in the MC. This commit replicate this hack (our
interfaces should be "actually" elastic) in order to make the
experimental backend produce valid RTL. Main changes are described
below.

1. Introduces a new Chisel LSQ module that instantiates an LSQ with
   the proper IO in these situations. The LSQ generator picks this
   models when the JSON config indicates that experimental generators
   should be picked (configuration introduced in a recent commit),
   and when the new `toMC` JSON key is set to true as well. Both the
   experimental and legacy backends set this key to an appropriate
   value to produce a correct LSQ in all cases.
2. Fixes the memory interface conversion logic in the `HandshakeTOHW`
   pass to take into consideration potential ports to other memory
   interfaces (both for MCs and LSQs).
3. Changes naming conventions for memory ports connecting an MC to an
   LSQ.