[Support] Move common system-interaction helpers to support library
This commit creates a new compile unit in the support library meant to
contain helpers for interacting with the system (e.g., executing
commands, path manipulation). A few system helpers that were previously
part of various tools are moved there to become usable by other parts of
the codebase. Included headers are modified as needed throughout the
codebase.