[HW] Move `PortNameGenerator` to dialect's library
The `PortNameGenerator` utility class, which derives port names for all
supported operations in the IR, was previously a private part of the
`HandshakeToHW` conversion pass and was therefore not usable by other
parts of the codebase. It may however be useful in different contexts,
and was therefore moved to the HW's dialect library.