[hls-verifier] Working testbench gen. for experimental backend
This improves the testbench generation logic to make it work with both
the legacy and experimental backend. The latter, as opposed to the
former, does not generate kernels with fake "start" nodes at the inputs
that are in charge of ensuring that a single token enters the circuit
for each function argument. It instead expects the testbench itself to
provide a single dataflow token to the circuit for each argument, which
it now does.

Once we get around to implementing our new circuit/memory interfaces the
circuits produced by Dynamatic will be able to internally control the
number of concurrent tokens inside of them (as they should). Therefore
this is a temporary (though sane) change to make both the legacy and
experimental backend work with the testbench generation logic.