[export-vhdl] Remove assumption on LSQs' name
The legacy backend expected LSQ interfaces and ports in its input DOT to
be capitalized, otherwise it would not recognize them. This initially
forced us to so some special casing in the DOT printer since we
typically use lowercase letters for naming components.

This commit removes this legacy backend limitation, allowing us to
remove this special-casing in the DOT printer.