[visual-dataflow] Add support for experimental backend
This commit adds the ability to visualize circuits generated by the
experimental backend in our interactive dataflow visualizer. This is
achieved through the introduction of a new (more reliable) C++
version of the wlf2csv script which converts ModelSim's waveform
output to a CSV file containing a list of state changes for all channels
in the circuits.

The `visualize` frontend command, similarly to others, now accepts an
`--experimental` flag to make it support circuits generated by the
experimental backend.