[Espresso] Algorithm for boolean simplification
This commit adds a modernized C/C++ implementation of the Espresso
algorithm to the repository. The original code was written at
Berkeley in 1989 and was modified throuhgout the years to port it to
newer C/C++ standards, on which we stack our own modifications to make
it work for modern C++ compilers and together with Dynamatic's
toolchain. Given the size of the implementation and its C roots, we
don't bother fixing stylistic issues and disable clang-tidy on the
subdirectories containing the implementation.

The Espresso algorithm is available as the `DynamaticSupportEspresso`
library in Dynamatic.

The implementation we started from came from the following repository:
https://github.com/classabbyamp/espresso-logic