[HandshakeToHW] Fix failure with conflicting module types
When two MLIR operations had different operand/result types but ended up
becoming instances of the same external hardware module, the pass would
fail because of conflicting IO types between the module and the hardware
instance corresponding to the MLIR operation converted last. This issue
arose for example with operations that, at the RTL level, do not care
for the data bus's actual content (e.g., buffers). If an integer and
floating point version of a specific buffer configuration existed
concurrently in the IR, the pass would fail. This is fixed by making the
pass's type converter replace floating point types with
identical-width integer types.

Going forward I think that using an `IntegerType` in HW to represent
everything that ends up as an opaque bitvector in RTL is the easiest and
semantically consistent way to do things.