[Support] Boolean logic library
This adds support for expressing and manipulating boolean expressions.
The library uses a tree structure to represent a boolean expression. It
supports functions such as: creating a boolean expressions directly
(through APIs) or by parsing a string, minimizing a boolean expression
using Espresso-II algorithm, and applying Shannon expansion to
transform a boolean expression into a tree of multiplexers (soon).