[Handshake] Add `handshake::BundleOp` and `handshake::UnbundleOp`
This introduces two new reciprocal operations to Handshake that allow to
go back and forth between "compound Handshake types"
(`!handshake.channel` and `!handshake.control`) and the individual
signals that make them up (making it possible to manipulate them
individually). These have rather complex (though almost identical)
verification logic to ensure that the conversion, whatever the
direction, is valid. Many new unit tests check that the implementation
produced an error for all possible types of invalid conversions.

The operations' assembly format is a bit verbose at the moment, this
should be improved in a future commit.