[Handshake] Add `handshake::ReshapeOp` operation
This introduces a new reshaping operation to Handshake which can change
the nature of a channel type's signals without altering the actual
payload carried by the channel, essentially performing an internal
rewiring of the various signals making up the channel. This is mostly
meant to be used during a pre-processing pass for the backend to
simplify channels around components that do not know how to handle or
care for extra signals.