[Handshake] Add Handshake version of Arith operations
This commit adds a Handshake version of a subset of the operations from
the `arith` dialect. As opposed to their counterparts from the built-in
dialect, these new version accept and return channel-typed values, which
we will need as soon as passes start using the new type system. Outside
the difference in expected types, these operations are equivalent to the
one they replace.