[Backedge] Use `PatternRewriter` to set edges when possible
When the `BackedgeBuilder` is created using a `PatternRewriter` instead
of an `OpBuilder`, all IR changes must go through the rewriter to ensure
that we do not run into undefined behavior. This was not the case in the
`Backedge::setValue` method, which directly replaced the edge's value
with the real one. `Backedge` can now be passed a rewriter during
construction to perform this task through it when needed.