[HandshakeToHW] Improve type conversion
This changes the type conversion function used throughout the
`HandshakeToHW` conversion pass. All sized types (and nested sized
types, as in a channel's data type for example) are now converted
to signless integers of the same width of the type they replace. At the
HW level we no longer care about the semantics of the values connecting
the circuit components. Converting everything to signless integers
ensures that we do not concretize duplicated modules which only differ
in the interpretation of their IO ports' content (which do not matter at
this point).

This also adds a couple unit tests for the pass to check for correct
type conversion behavior.