[Handshake] Better assembly format for (un)bundle
The assembly format of the `handshake::BundleOp` and
`handshake::UnbundleOp` operations was very verbose; in particular iot
contained a lot of redundancy in the types of all operands/results, when
a single control/channel type is in fact enough to encode the type of
all operands/results.

This commit implements custom (similar) printers/parsers for both
operations, which consequentially simplify the ops' printed format and
additionally reduces the number of potential issues with the ops' text
serialization (hence the removal of some unit tests, in addition to the
modification of others).

Fixes #99.