[HandshakeReshapeChannels] Simplify channels before emission
Introduces a new pass meant to be a last Handshake-level pre-processing
step before lowering to HW and eventually to RTL. The pass simplifies
the form of channel-typed values around operations that do not read the
content of extra signals (and potentially the data signal). This relies
on

- eligible operations implementing the `ReshapableChannelsInterface`
  interface to signify to the pass which of their operand channels may
  be reshaped and how, and
- the insertion of `handshake::ReshapeOp` in the IR to go back-and-forth
  between complex and simple channel forms.

New unit tests are added for the pass.