[RTL] Refactor typing system for RTL parameters
This improves the internal representation of constrained RTL parameter
types, making it easier to define new types in the future.

While the existing system was working it was difficult to understand
what exactly one should do to define a new RTL parameter type. Some
methods which were semantically static were in fact pure virtual methods
of an abstract base type, it was unclear who was responsible for
deserializing JSON data into a particular set of constraints, etc.

The `RTLType` type, which is no longer abstract, now relies on a mix of
CRTP and concept-based polymorphism to provide a clean non-templated API
to client code while allowing concrete RTL parameter types and
constraints to have a leaner and more meaninful API than before.