[Backend] Support for `handshake.join` operations
While we already have a VHDL join component in our backend config, it is
only used internally as part of other components and not actually linked
to the `handshake.join` operation (which we never insert in the IR as
part of the current compilation flow), which therefore cannot currently
be lowered.

This adds support for lowering such operations to RTL. The corresponding
VHDL component is duplicated almost as is with two extras clk and rst
ports to satisfy the current (unsatisfactory) Handshake to HW
conversion's behavior which always adds clock and reset ports to
Handshake components. The new component simply instantiates the old one
(now part of the "support" components) without the clk and rst. This is
not a good way of doing this, and will be fixed when the conversion pass
gets a refactoring round.