[JSON] Improve custom JSON object mapper
The recently introduced `OptionalObjectMapper` was already useful but
has some limitations that this commit lifts.

1. Instances of the type have to be built from a JSON object while it is
   more common in LLVM's JSON deserialization methods to pass around a
   JSON value instead (which can be converted to an object, if it is
   one). The type can now additionally be built from a JSON value.
2. As its name indicate, the type only supports optional mappings. It
   now also supports non-optional mappings, which, upon a missing key,
   permanently mark the mapping state as invalid.

The type is renamed from `OptionalObjectMapper` to `ObjectDeserializer`.

This allows us to refactor the `fromJSON` methods on RTL constraints
into free LLVM-standard-compliant functions that take a JSON value as
first argument and the type to deserialize into as second argument.