[Handshake] Move attribute def/decl to separate compile unit
This moves all declarations and definitions of custom Handshake
attributes to a dedicated compile unit in the Handshake library, as is
standard in MLIR. Include paths are updated to reflect these changes.

A few storage data-structures for these attributes are also not in the
right namespace (directly under "dynamatic" instead of
"dynamatic::handshake"); this fixes that as well and updates the code
using them wherever necessary.