[Backend] Fix inconsistencies in generic parameter names
There are some inconsistencies in the name of the generic RTL parameters
declared in the RTL configuration/`HandshakeToHW` conversion pass and
the actual VHDL implementations. This is not a problem when emitting the
top-level RTL in VHDL because there we do not rely on generic parameter
names (only their order). However, when emitting the top-level RTL in
Verilog we reference generic parameters by name so they need to match.

This commit ensures that this is the case for all of our components.