[Handshake] Delete `ReturnOp` operation
The `ReturnOp` operations, which we use to feed the terminator of
Handshake functions, is equivalent to a TEHB (it is converted into one
during Handshake to HW lowering) and, generally speaking, serves no
purpose. This commit deletes it from the Handshake dialect and removes
references to it throughout the codebase. In particular, most unit tests
need mechanical changes to reflect the operation's deletion.