[Handshake] Only master mem. ifaces. produce completion signal
Every memory interface (MC and LSQ) currently produces a completion
signal as its last result that goes to the `handshake::EndOp` Handshake
function terminator. In cases where two memory interfaces reference the
same memory region this means that two completion signals are generated
for the memory region, which does not make semantic sense.

This commit introduces the notion of master/slave interface and defines
new conventions for how each connect to the function terminator. Only
master interfaces produce completion signals, and there must only be a
single master interface for each memory region (the number of slave
interfaces for any given region is unrestricted). Only our LSQ interface
can ever be a slave interface currently, so it no longer produces a
completion signal when it is connected to an MC, which serves as the
master.

The change require light modifications (which largely simplify the code)
in the CF to Handshake and Handshake to HW lowerings as well as in the
LSQ minimization pass.