[LSQ] Removed legacy LSQ gen. support
Now that the legacy backend is gone, there is no need to keep around
the legacy-compatible LSQ BRAM Chisel component. This commit thus
removes it.

The LSQ JSON configuration file no longer expects the "experimental"
key since the experimental backend is now the only one present. We
also rename the two LSQ Chisel component to follow the master/slave
interface nomenclature introduced in 5804d2f. The "toMC" key in the
JSON config becomes "master", which should be `true` when a master
LSQ interface should be generated.