[CFG] Do not emit error on BB control id. failure
Our CFG analysis emits an error when it cannot identify the control
value of one of the "basic block" of a Handshake function. This is
overkill; client code may simply wish to skip an optimization in the
case where this identification is unsuccessful. This commit makes the
method simply return a logical failure without producing an erorr.

Fixes #110.