Circuit interface redesign
This large contribution makes Dynamatic implement our new and improved
circuit interface, as proposed in `docs/Specs/CircuitInterface.md`. In
addition to the regular function arguments and results inherited from C,
each of our circuits now has a control-only `start` input and `end`
output, and a similar pair of ports for each memory region, respectively
to signal that memory accesses can start being issued to an external
memory and to signal to the outside world that no more memory accesses
will be issued by the circuit for a particular memory region.

This requires significant changes in our memory interfaces, both in MLIR
and in RTL.

1. In MLIR, every master memory interface gets two extra operands, one
   for the memory's "start" signal coming from the circuit's boundary
   and one coming from the circuit itself to signal that it has made its
   last control flow decision and therefore that no more accesses than
   the ones already announced will come. This helps generate the
   memory's "end" signal going to the circuit's boundary.
2. In RTL (VHDL only at this point), necessary changes are made to the
   MC's and LSQ's architecture to implement the new semantics using
   these new ports. The logic seems to work but is probably incomplete,
   a good RTL designer should probably take a look at it to verify that
   it really makes sense.

Notably, this redesign removes the need for the `end` component in RTL.
When lowering to HW, the `handshake::EndOp` Handshake terminator no
longer translates to a component, instead its operands directly get fed
to the `hw::OutputOp` terminator.

Significant changes are made to the testbench generation logic to
support this new interface, as a result our testbenches are a bit more
readable.

As a side effect, fixes #118