[LSQ] Fix master/slave identification logic
The logic that attempts to determine whether the LSQ is a slave
interface connected to an MC is broken; it never detects a connected MC
even if there is one. This fixes it by correctly looking for the user of
the last LSQ result, which is the store data bus going to the MC when
the LSQ is a slave interface.

Fixes #123.