[Verilog] Add formatter config. for Verilog
This commit adds a configuration file for the `verible-verilog-format`
tool, which is the official formatter of the Verible developer
toolchain. The configuration file need to be explicitly passed to the
tool to be picked up.

```
verible-verilog-format --flagfile .verilog-format <input file>
```

An extension to run the tool automatically in editor exists for VSCode:
https://marketplace.visualstudio.com/items?itemName=mshr-h.VerilogHDL