[handshake-simulator] Move to experimental Support (#128)
This moves all entities previously part of the Handshake simulator tool
into their own dedicated compile unit in the experimental "Support"
library. This allows other part of the codebase to run the simulator.
The skeleton of a test pass for the Handshake simulator is included as
an example and will be filled with actual test cases in future commits.