[RTL] `RTLChannelType` -> `RTLDataflowType`
The `RTLChannelType` can currently only deserialize dataful channels
(`handshake::ChannelType` in the IR). As such there is no current way to
encode a control-only channel (`handshake::ControlType`) as an RTL
parameter.

This commit adds support for the latter, allowing both dataflow-like
types to be represented using the same RTL parameter type. The same
constraining logic may be used for control-only channels (which are
simply considered like channels with 0-bit data and no extra signals).
The RTL parameter type's name is changed to `RTLDataflowType` to carry
the meaning that all dataflow-like types are supported.

Unit tests for the `RTLChannelType` are modifed to make sure that
control-only channels can be deserialized correctly.