[Backend] Use "dataflow" RTL type parameter everywhere
While there is already support for encoding dataflow-like types in RTL
parameters, the backend does not yet actually use this support anywhere.
Instead, the `HandshakeToHW` pass encodes the data width of
dataflow-like types as an "unsigned" RTL parameter, which works but
cannot represent the full range of channel types our new type system can
represent.

This commit finally fixes that and updates the lowering logic and RTL
configurations to use the "dataflow" type instead everywhere where it is
relevant. While our actual component implementations do not yet support
any "complex channel type" (e.g., with extra signals), this is now
properly encoded in the RTL configs and can be easily modified in the
future as needed. Most components which use a parameter name ending in
"_WIDTH" see these parameters now ending in "_TYPE" to reflect the fact
that the parameter now encodes a full dataflow type and no longer just a
data width. This creates mechanical renamings in most of our RTL
components.