[HandshakeToHW] Fix float constant encoding to binary
The logic to convert a floating constant value to its IEEE 754 binary
representation is faulty (it produces a string of 0s in most cases if
not all the time). This results in incorrectly encoded constant values
during RTL emission and simulation errors down the line.

This commit fixes that by using LLVM's bit-by-bit `APFloat` to `APInt`
conversion, followed by a `std::bitset` of the correct width.

Fixes #122.