[Attribute] Rework concept of "unique attributes"
The Dynamatic attribute helpers offer functions to work on "unique
attributes" i.e., discardable attributes whose name is given by the
mnemonic of their type. The concept is however flawed, since it doesn't
honor MLIR's rule that discardable attributes should be prefixed by a
dialect name.

This commit replaces the concept of "unique attributes" with that of
"dialect attributes". These have the same property of including the
attribute type's mnemonic in their name, except that now this name is
prefixed by a dialect name, honoring MLIR's rule. Functions that worked
with "unique attributes" are modified and renamed to reflect this
concept change. The implementation of unctions working on operand
attributes are also modified so that their associated container
attribute now behaves as a dialect attribute as well.

Unit tests are updated wherever necessary. This amounts to adding a
"handshake." prefix to the name of some existing attributes.