[Attribute] Use MLIR methods on discardable attribute
The helper functions to work with "dialect attributes" are using the
generic MLIR methods for working with attributes (inherent or
discardable). However, "dialect attributes" are by definition
discardable attributes so it make mores sense to use the version of
these methods that work exclusively on discardable attributes.

This commit does just that. This should marginally increase
performance and decrease the probability of user error when working with
dialect attributes. At the same time, a function to remove a dialect
attribute is introduced.