[Handshake] Modularize LSQ minimization flow
The current LSQ minimization pass (`HandshakeMinimizeLSQUsage`) performs
two sequential tasks to achieve its goal of reducing the size and number
of LSQs in the circuit.

1. It analyzes the structure of the circuits to identify load and store
   ports that currently connect to an LSQ but do not need to.
2. It uses the results of the previous analysis to modify memory
   interfaces in the circuit so that only the memory access ports that
   are strictly required to connect to the LSQ actually do.

Modifying memory interfaces in step 2 requires the identification of
the control value of each basic block in the Handshake function under
consideration. Due to our limited heuristic for doing so (part of the
CFG support), this can sometimes fail, which prevents the pass from
applying the optimizations it identified in step 1. As such it becomes
unclear in the resulting IR whether the presence of an LSQ that can
seemingly be reduced/removed is due to a fault in the structural circuit
analysis of step 1 or to a limitation of our CFG analysis in step 2.

This commit makes such faults clearer by splitting the pass into two in
alignment with the two steps.

1. The `HandshakeMinimizeLSQUsage` pass becomes
   `HandshakeAnalyzeLSQUsage` and only performs the first step described
   above. It communicates the results of its analysis to the second pass
   using `handshake::MemoryInterfaceAttr` attributes attached to all
   memory access ports it encountered.
2. We introduce a new `HandshakeReplaceMemoryInterfaces` pass that is
   now responsible for the second step of this transformation: replacing
   existing memory interfaces with new ones according to the attributes
   left by the LSQ analysis pass (or to any other pass for that matter).
   This pass fails when our control-value-identifying heuristic cannot
   determine the control value of a basic block, making this type of
   error clearly understandable by the user.

In this modularization process, this commit makes light changes to the
memory interface infrastructure which reverberate in some other files
and passes.