Skeleton of FTD `cf` to `handshake` conversion pass
The commit implements the skeleton for the new experimental pass implementing FPGA'22 and FPGA'23.
The skeleton is now filled with the same operations performed by the "official" CF to handshake conversion.
To make it work, the declarations of some structures have been moved from `CfToHandshake.cpp` to `CfToHandhshake.h`, so that they are visible from the experimental folder as well.
A new flag has been added to the `compile` command, in order to use this conversion pass instead of the normal one.
To use it, run `compile --fast-token-delivery`.