[Handshake] Fix incorrect mem. iface. port identifiation
The logic to determine the range of operands and results of a memory
interface that represent ports to other memory interfaces is incorrect.
In particular, it may consider the control start and control end
operands of a memory interface as part of the operands to other memory
interfaces. This commit fixes that, which in particular resolves
warnings we would sometimes get during buffer placement due to these
ports being incorrectly marked as unbufferizable.

Fixes #124.