[DOTPrinter] Print port names and indices
This makes the DOT printer write, for each edge in the graph, the source
and destination port names of the two operations that it connects. These
are derived using the same logic as in the `HandshakeToHW` conversion
pass. The helper class that caches all port names for a given operation
is moved from the HW dialect's library to the Handshake dialect's
library, which makes more sense since port names for most operations are
expressed through the `handshake::NamedIOInterface`.