[FTD][ControlDependenceAnalysis] Added CDA anaylsis library
This commit adds the CDA analysis library as designed in commmit `32ad8d23727e62e8667df2485becb8b3ce9f755a`.
The main modification is ubout the usage of const functions within the library.
Since `matchAndRewrite` in `LowerFuncToHandshake` is `const`, then all the methods used on its attributes must be const as well.
However, this does not limit the functionalities of the library, as all the public methods are in charge of performing loop-ups - withuot any update of the attributes.