[FTD][CfToHandshake] Load-Store Queue Allocation Network - Groups
This commits implements the logic to create the groups' graph for the LSQ allocation network.
Most of the code for the allocation was taken from `8fd166e1162dcbd6397d4e61107ff2a404db3fe5`.
Minor refactoring has been done to merge the code with the current status of the repo.