[FTD][ControlDependenceAnalysis] Update key type in dependencies map
In the previous version of the library, the information about the dependencies were addressed using instances of `func::FuncOp`.
Using this type as key is no longer possible once that the conversion pass lowers it to `handshake::FuncOp`, while we are still interested on the basic blocks information.
For this reason, functions are now addressed using their name, which is maintained throughout the lowering of the type.
Moreover, from a front-end point of view, it is not possible to have two functions with the same name, so there is no conflict in addressing the dependencies.