[FTD][ControlDependenceAnalaysy] Modified analysis structure
The analysis can be now applied both to functions of type `func::FuncOp` and `handshake::FuncOp`.
It is supposed to be called whenever it is needed on the required object, rather than connected to the conversion pass.
In this way, the structure can be used in different scenarios without compromising generality.